Fix to #11831 - Query: compilation error for queries with groupby anonymous object and table splitting
When remapping member infos in a table splitting scenario we were not taking into the account that the element could be not a MemberInfo (it was a NullConditionalExpression because nav rewrite added null protection to it). However in case of accessing a property of an IGrouping, it's not necessary, because grouping can never be null.

Fix is to check for presence of GroupResultOperator when determining whether a given member access needs null compensation.